% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqLogo.R
\docType{data}
\name{alphabet.dna.gap}
\alias{alphabet.dna.gap}
\title{DNA alphabet with gaps}
\format{
An object of class \code{list} of length 2.
}
\usage{
alphabet.dna.gap
}
\description{
DNA alphabet with gaps
}
\keyword{datasets}
