% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqLogo.R
\docType{data}
\name{alphabet.protein.gap}
\alias{alphabet.protein.gap}
\title{Amino acid alphabet with gaps}
\format{An object of class \code{list} of length 2.}
\usage{
alphabet.protein.gap
}
\description{
Amino acid alphabet with gaps
}
\keyword{datasets}
