% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqLogo.R
\docType{data}
\name{alphabet.protein}
\alias{alphabet.protein}
\title{Amino acid alphabet}
\format{An object of class \code{list} of length 2.}
\usage{
alphabet.protein
}
\description{
Amino acid alphabet
}
\keyword{datasets}
