% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqLogo.R
\docType{data}
\name{alphabet.rna.gap}
\alias{alphabet.rna.gap}
\title{RNA alphabet with gaps}
\format{An object of class \code{list} of length 2.}
\usage{
alphabet.rna.gap
}
\description{
RNA alphabet with gaps
}
\keyword{datasets}
