% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepLogoR.R
\name{filter.by.gaps}
\alias{filter.by.gaps}
\title{Filters columns (sequence positions) by gaps}
\usage{
filter.by.gaps(percent.gap)
}
\arguments{
\item{percent.gap}{the maximum fraction of gaps allowed to retain a column}
}
\value{
function that, given a \link{DLData} object, returns \code{TRUE} 
  for every column that does not exceed the specified number of gaps
}
\description{
Filters columns (sequence positions) by gaps
}
\examples{
fun <- filter.by.gaps(percent.gap = 0.1)
}
\author{
Jan Grau <grau@informatik.uni-halle.de>
}
