% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepLogoR.R
\name{revcom}
\alias{revcom}
\title{Reverse complement}
\usage{
revcom(data)
}
\arguments{
\item{data}{the data}
}
\value{
the reverse complement
}
\description{
Determine the reverse complementary \code{DLData} object. Only works for DNA or RNA. Data may include gap symbols.
}
\examples{
data <- DLData(c("ACGT", "ATTA"))
revcom(data)
}
\author{
Jan Grau <grau@informatik.uni-halle.de>
}
